feat: ful implementation of `mark as read` for account, product and individual notifications
