--- src/utils/notifications.ts
@@ -1,227 +1,0 @@
-import log from 'electron-log';
-import type {
-  Account,
-  AccountNotifications,
-  AtlassifyNotification,
-  AtlassifyState,
-  SettingsState,
-} from '../types';
-import { getNotificationsForUser } from './api/client';
-import { determineFailureType } from './api/errors';
-import type { AtlassianNotification, Category, ReadState } from './api/types';
-import { getAccountUUID } from './auth/utils';
-import { hideWindow, showWindow, updateTrayIcon } from './comms';
-import { Constants } from './constants';
-import { READ_STATES } from './filters';
-import { openNotification } from './links';
-import { isWindows } from './platform';
-import { getAtlassianProduct } from './products';
-
-export function setTrayIconColor(notifications: AccountNotifications[]) {
-  const allNotificationsCount = getNotificationCount(notifications);
-
-  updateTrayIcon(allNotificationsCount);
-}
-
-export function getNotificationCount(notifications: AccountNotifications[]) {
-  return notifications.reduce(
-    (memo, acc) => memo + acc.notifications.length,
-    0,
-  );
-}
-
-export function hasMoreNotifications(notifications: AccountNotifications[]) {
-  return notifications?.some((n) => n.hasNextPage);
-}
-
-export const triggerNativeNotifications = (
-  previousNotifications: AccountNotifications[],
-  newNotifications: AccountNotifications[],
-  state: AtlassifyState,
-) => {
-  const diffNotifications = newNotifications
-    .map((accountNotifications) => {
-      const accountPreviousNotifications = previousNotifications.find(
-        (item) =>
-          getAccountUUID(item.account) ===
-          getAccountUUID(accountNotifications.account),
-      );
-
-      if (!accountPreviousNotifications) {
-        return accountNotifications.notifications;
-      }
-
-      const accountPreviousNotificationsIds =
-        accountPreviousNotifications.notifications.map((item) => item.id);
-
-      const accountNewNotifications = accountNotifications.notifications.filter(
-        (item) => {
-          return !accountPreviousNotificationsIds.includes(`${item.id}`);
-        },
-      );
-
-      return accountNewNotifications;
-    })
-    .reduce((acc, val) => acc.concat(val), []);
-
-  setTrayIconColor(newNotifications);
-
-  // If there are no new notifications just stop there
-  if (!diffNotifications.length) {
-    return;
-  }
-
-  if (state.settings.playSoundNewNotifications) {
-    raiseSoundNotification();
-  }
-
-  if (state.settings.showSystemNotifications) {
-    raiseNativeNotification(diffNotifications);
-  }
-};
-
-export const raiseNativeNotification = (
-  notifications: AtlassifyNotification[],
-) => {
-  let title: string;
-  let body: string;
-
-  if (notifications.length === 1) {
-    const notification = notifications[0];
-    title = isWindows() ? '' : notification.title;
-    body = notification.entity.title; // TODO Confirm this mapping
-  } else {
-    title = 'Atlassify';
-    body = `You have ${notifications.length} notifications.`;
-  }
-
-  const nativeNotification = new Notification(title, {
-    body,
-    silent: true,
-  });
-
-  nativeNotification.onclick = () => {
-    if (notifications.length === 1) {
-      hideWindow();
-      openNotification(notifications[0]);
-    } else {
-      showWindow();
-    }
-  };
-
-  return nativeNotification;
-};
-
-export const raiseSoundNotification = () => {
-  const audio = new Audio('../../assets/sounds/notification.wav');
-  audio.volume = 0.2;
-  audio.play();
-};
-
-function getNotifications(state: AtlassifyState) {
-  return state.auth.accounts.map((account) => {
-    return {
-      account,
-      notifications: getNotificationsForUser(account, state.settings),
-    };
-  });
-}
-
-export async function getAllNotifications(
-  state: AtlassifyState,
-): Promise<AccountNotifications[]> {
-  const responses = await Promise.all([...getNotifications(state)]);
-
-  const notifications: AccountNotifications[] = await Promise.all(
-    responses
-      .filter((response) => !!response)
-      .map(async (accountNotifications) => {
-        try {
-          const res = (await accountNotifications.notifications).data;
-
-          const rawNotifications =
-            res.data.notifications.notificationFeed.nodes;
-
-          let notifications = mapAtlassianNotificationsToAtlassifyNotifications(
-            accountNotifications.account,
-            rawNotifications,
-          );
-
-          notifications = filterNotifications(notifications, state.settings);
-
-          return {
-            account: accountNotifications.account,
-            notifications: notifications,
-            hasNextPage:
-              res.extensions.notifications.response_info.responseSize ===
-              Constants.MAX_NOTIFICATIONS_PER_ACCOUNT, // TODO - there is a bug in the Atlassian GraphQL response
-            error: null,
-          };
-        } catch (error) {
-          log.error(
-            'Error occurred while fetching account notifications',
-            error,
-          );
-          return {
-            account: accountNotifications.account,
-            notifications: [],
-            hasNextPage: false,
-            error: determineFailureType(error),
-          };
-        }
-      }),
-  );
-
-  return notifications;
-}
-
-export function mapAtlassianNotificationsToAtlassifyNotifications(
-  account: Account,
-  notifications: AtlassianNotification[],
-): AtlassifyNotification[] {
-  return notifications?.map((notification: AtlassianNotification) => ({
-    // TODO Improve the fidelity of this mapping
-    id: notification.headNotification.notificationId,
-    title: notification.headNotification.content.message,
-    readState: notification.headNotification.readState as ReadState,
-    unread: notification.headNotification.readState === READ_STATES.unread.name,
-    updated_at: notification.headNotification.timestamp,
-    url: notification.headNotification.content.url,
-    path: notification.headNotification.content.path[0],
-    entity: notification.headNotification.content.entity,
-    category: notification.headNotification.category as Category,
-    actor: notification.headNotification.content.actor,
-    product: getAtlassianProduct(notification),
-    account: account,
-  }));
-}
-
-export function filterNotifications(
-  notifications: AtlassifyNotification[],
-  settings: SettingsState,
-): AtlassifyNotification[] {
-  return notifications.filter((notification) => {
-    if (
-      settings.filterCategories.length > 0 &&
-      !settings.filterCategories.includes(notification.category)
-    ) {
-      return false;
-    }
-
-    if (
-      settings.filterReadStates.length > 0 &&
-      !settings.filterReadStates.includes(notification.readState)
-    ) {
-      return false;
-    }
-
-    if (
-      settings.filterProducts.length > 0 &&
-      !settings.filterProducts.includes(notification.product.name)
-    ) {
-      return false;
-    }
-
-    return true;
-  });
-}
